Internal Classes to internal
