+++ Fody/Fody.csproj
@@ -69,12 +69,15 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Configuration.cs" />
-    <Compile Include="EnumerableEx.cs" />
-    <Compile Include="FormatStringTokenResolver.cs" />
-    <Compile Include="LibGitEx.cs" />
+    <Compile Include="Internal\EnumerableEx.cs" />
+    <Compile Include="Internal\FormatStringTokenResolver.cs" />
+    <Compile Include="Internal\LibGitEx.cs" />
     <Compile Include="ModuleWeaver.cs" />
-    <Compile Include="..\CommonAssemblyInfo.cs" />
+    <Compile Include="..\CommonAssemblyInfo.cs">
+      <Link>Properties\CommonAssemblyInfo.cs</Link>
+    </Compile>
     <Compile Include="PeImage\BinaryReaderExtensions.cs" />
     <Compile Include="PeImage\BinaryWriterExtensions.cs" />
     <Compile Include="PeImage\DllCharacteristicsType.cs" />
