+++ Fody/Internal/EnumerableEx.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Stamp.Fody.Internal
+{
+    public static class EnumerableEx
+    {
+	
+        public static bool IsEmpty<T>(this IEnumerable<T> enumerable)
+        {
+            return !enumerable.Any();
+        }
+    }
+}
